+++ resources/views/livewire/attendance/poll-edit.blade.php
@@ -1,5 +1,6 @@
 @php
     /** @var \App\Livewire\Forms\PollForm $pollForm */
+    $hasPollShowPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\AttendancePoll::ATTENDANCE_POLL_SHOW_PERMISSION) ?? false;
     $hasAttendanceShowPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Attendance::ATTENDANCE_SHOW_PERMISSION) ?? false;
     $hasAttendanceEditPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Attendance::ATTENDANCE_EDIT_PERMISSION) ?? false;
 @endphp
@@ -14,27 +15,27 @@ addEvents() {
     $wire.addEventsToSelection(this.additionalEventList);
     this.additionalEventList = [];
 }}">
-    <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg mb-5 p-5 flex justify-between gap-2 items-center">
-        <button type="button"
-                x-data="{ clickCnt: 0, onClick() {
-                        if(this.clickCnt > 0) {
-                            $wire.deletePoll();
-                        } else {
-                            this.clickCnt++;
-                            $el.innerHTML = 'Are you sure?';
-                        }
-                    }}"
-                x-on:click="onClick()" title="Delete this poll"
-                class="btn btn-danger">{{ __('Delete poll') }}</button>
-        <div class="flex gap-2 items-center">
-            <span x-cloak class="text-gray-500 text-xs mt-1"
-                  x-show="additionalEventList.length > 0">Add selected events or unselect them.</span>
-            <div class="ml-auto">
-                <button type="button" class="btn btn-primary inline-flex" wire:click="savePoll"
+    <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg mb-5 p-5 flex justify-end gap-2 items-center">
+         <span x-cloak class="text-gray-500 text-xs mt-1"
+               x-show="additionalEventList.length > 0">Add selected events or unselect them.</span>
+        <x-button-dropdown>
+            <x-slot name="mainButton">
+                <button type="button" class="btn-success p-2 text-xs inline-flex items-center gap-2"
                         x-bind:disabled="additionalEventList.length > 0"
-                        title="Update attendance poll">{{ __('Save') }}</button>
-            </div>
-        </div>
+                        wire:click="savePoll" title="Delete this poll">
+                    <i class="fa-solid fa-floppy-disk"></i> {{ __('Save') }}</button>
+            </x-slot>
+            @if($hasPollShowPermission)
+                <a href="{{route('attendancePoll.statistic', $pollForm->poll->id)}}" class="btn-secondary text-xs p-2 inline-flex"
+                   title="Show summary of attendance poll">
+                    {{__("Show summary")}}
+                </a>
+            @endif
+            <button type="button" class="text-xs p-2 btn-danger inline-flex gap-2"
+                    wire:confirm="{{__('Are you sure you want to delete this attendance poll?')}}"
+                    wire:click="deletePoll" title="Delete this attendance poll."
+            ><i class="fa-solid fa-trash"></i> {{ __('Delete event') }}</button>
+        </x-button-dropdown>
     </div>